requirements.txt
langgraph
langchain
langchain-core
langchain-community
python-dotenv
langchain-groq
duckduckgo-search
dateparser
json
requests
regex